fix(ecDashboard): code fixes
